Add syscalls crate
