+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
   "crypto/crypto",
   "crypto/math",
   "bin/cli",
+  "syscalls",
 ]
 
 resolver = "2"
