+++ executor/programs/rust/panic/Cargo.toml
@@ -6,5 +6,4 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-embedded-alloc = "0.6"
-riscv = { version = "0.15", features = ["critical-section-single-hart"] }
+lambda-vm-syscalls = { path = "../../../../syscalls" }
