+++ executor/programs/rust/print/Cargo.toml
@@ -6,3 +6,4 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+lambda-vm-syscalls = { path = "../../../../syscalls" }
