+++ syscalls/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "lambda-vm-syscalls"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+embedded-alloc = "0.6"
+riscv = { version = "0.15", features = ["critical-section-single-hart"] }
